requirements.txt
Pillow
openpyxl
python-pptx
tk
